+++ Sprites/Characters/Enemy/Knight.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Project.Characters;
+using Project.Enums;
+using Project.Interfaces;
 using Project.Managers;
 using Project.UI;
 using System;
@@ -10,11 +12,19 @@ namespace Project.Sprites.Characters.Enemy
 {
     public class Knight : Enemy
     {
+        private KnightSword weapon;
         public Knight(Texture2D texture, Texture2D heartTexture) : base(texture, heartTexture)
         {
             Position = new Vector2(48, 1);
             Speed = new Vector2(2, 1);
             Health = new Health(3, heartTexture);
+            Damage = 1;
+            weapon = new KnightSword(Game1.swordTexture, this);
+        }
+
+        public override void Attack(GameTime gameTime)
+        {
+            weapon.Update(gameTime);
         }
 
         internal override void Move()
@@ -27,7 +37,7 @@ namespace Project.Sprites.Characters.Enemy
 
             Vector2 targetPosition = distanceToLeft < distanceToRight ? playerLeft : playerRight;
             Vector2 direction = targetPosition - Position;
-            
+
             if (direction.Length() < 4)
             {
                 direction = Vector2.Zero;
@@ -39,7 +49,15 @@ namespace Project.Sprites.Characters.Enemy
             }
 
 
-            MovementManager.Move(this, animationState, direction);
+            CharacterAnimation animation;
+            MovementManager.Move(this, animationState, direction, out animation);
+            Direction = animation;
+        }
+
+        public override void Draw(SpriteBatch spriteBatch)
+        {
+            weapon.Draw(spriteBatch);
+            base.Draw(spriteBatch);
         }
     }
 }